#13 프로젝트 구조를 변경하였다.
