--- api-module/src/main/java/com/example/ApiModuleApplication.java
@@ -1,13 +1,0 @@
-package com.example;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ApiModuleApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(ApiModuleApplication.class, args);
-    }
-
-}
