--- api-module/src/main/java/com/example/apimodule/dto/DuplicateCheckOwnerEmailRequestDto.java
@@ -1,13 +1,0 @@
-package com.example.apimodule.dto;
-
-import lombok.Getter;
-
-@Getter
-public class DuplicateCheckOwnerEmailRequestDto {
-
-    private String ownerEmail;
-
-    public OwnerDto toDto() {
-        return OwnerDto.builder().ownerEmail(ownerEmail).build();
-    }
-}
