+++ boot/owner-api/src/main/java/com/example/dev/dto/DuplicateCheckOwnerEmailRequestDto.java
@@ -1,0 +1,13 @@
+package com.example.dev.dto;
+
+import lombok.Getter;
+
+@Getter
+public class DuplicateCheckOwnerEmailRequestDto {
+
+    private String ownerEmail;
+
+    public OwnerDto toDto() {
+        return OwnerDto.builder().ownerEmail(ownerEmail).build();
+    }
+}
