+++ data/core/build.gradle
@@ -1,0 +1,4 @@
+
+dependencies {
+    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+}
