+++ httpclient/request.http
@@ -1,0 +1,4 @@
+GET http://localhost:8888/?name=jisu
+
+HTTP/1.1
+Content-Type: application/json
