--- module-core/src/main/java/com/example/qtable/entity/AuditingFields.java
@@ -1,25 +1,0 @@
-package com.example.qtable.entity;
-
-import lombok.Getter;
-import org.hibernate.annotations.ColumnDefault;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.MappedSuperclass;
-
-import java.time.LocalDateTime;
-
-@Getter
-@EntityListeners(AuditingEntityListener.class)
-@MappedSuperclass
-public abstract class AuditingFields {
-
-    @ColumnDefault("'N'")
-    private String deleteAt;
-    @CreatedDate
-    private LocalDateTime createdAt;
-    @LastModifiedDate
-    private LocalDateTime updatedAt;
-}
