--- module-core/src/main/resources/application.yml
@@ -1,20 +1,0 @@
-spring:
-  main:
-    allow-bean-definition-overriding: true
-  datasource:
-    url: jdbc:h2:mem:testdb
-    username: sa
-    password:
-    driver-class-name: org.h2.Driver
-  jpa:
-    database: h2
-    database-platform: org.hibernate.dialect.H2Dialect
-    hibernate:
-      ddl-auto: create # or `create`
-    properties:
-      hibernate:
-        show_sql: true # to System OutDataSourceConfiguration
-        format_sql: true
-        use_sql_comments: true
-        jdbc:
-          time_zone: Asia/Seoul # @CreatedDate ??? Time Zone ? ??
