--- module-core/src/test/java/com/example/qtable/QtableApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.qtable;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class QtableApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
